Backend/requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
audioread==3.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
decorator==5.2.1
Django==4.2.28
django-cors-headers==4.9.0
django-environ==0.13.0
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.16.1
dotenv==0.9.9
filelock==3.25.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
numba==0.64.0
numpy==2.4.2
packaging==26.0
platformdirs==4.9.2
pooch==1.9.0
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydub==0.25.1
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.2.1
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
service-identity==24.2.0
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typer==0.24.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
